Migrate activity logic out to new representatives
