+++ app/src/main/kotlin/org/pointyware/accountability/viewer/ViewerActivity.kt
@@ -115,70 +115,4 @@ class ViewerActivity : AppCompatActivity() {
     }
 
     // endregion
-
-    private fun configureCallButtons() {
-        // TODO: migrate sync properties to StateFlow; bind button visibility to StateFlow
-        // setup preview surface
-        if (viewModel.callingEnabled) {
-            Timber.v("Calling Enabled")
-
-            previewCallButton.visibility = if (viewModel.friendEnabled) {
-                Timber.v("Friend Enabled")
-
-                previewCallButton.setOnClickListener {
-                    startFriendlyCall()
-                }
-                View.VISIBLE
-            } else { View.GONE }
-
-            policeCallButton.visibility = if (viewModel.policeEnabled) {
-                Timber.v("Emergency Enabled")
-
-                policeCallButton.setOnClickListener {
-                    startEmergencyCall()
-                }
-                View.VISIBLE
-            } else { View.GONE }
-
-        } else {
-            Timber.v("Calling Disabled")
-
-            previewCallButton.visibility = View.GONE
-            policeCallButton.visibility = View.GONE
-        }
-    }
-
-    private fun startFriendlyCall() {
-        Timber.v("Calling Friend")
-
-        viewModel.contactNumber?.also {
-
-            startCall(true, it)
-
-        } ?: run {
-
-            // if no number saved, open dialer
-            startCall(false)
-        }
-    }
-
-    private fun startEmergencyCall() {
-        Timber.v("Dialing Emergency Services")
-
-        startCall(false, viewModel.policeNumber)
-    }
-
-    @Deprecated("Use StartCallUseCase instead",
-        ReplaceWith("StartCallUseCase.invoke(callAfterDial, number)"))
-    private fun startCall(callAfterDial: Boolean, number: String = "") {
-        val action = if (callAfterDial) { Intent.ACTION_CALL } else { Intent.ACTION_DIAL }
-        val intent = Intent(action).apply {
-            data = Uri.parse("tel:$number")
-        }
-        try {
-            startActivity(intent)
-        } catch (exception: ActivityNotFoundException) {
-            Toast.makeText(baseContext, "Activity Could Not Be Found to Handle Calling", Toast.LENGTH_LONG).show()
-        }
-    }
 }
